+++ AffiseAttributionLib/Classes/modules/advertising/AffiseAdvertising.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+internal class AffiseAdvertising : NSObject, AffiseModuleApiWrapper {
+    typealias API = AffiseAdvertisingApi
+    var api: API?
+    var module: AffiseModules = .Advertising
+}
+
+extension AffiseAdvertising : AffiseModuleAdvertisingApi {
+    
+    func hasModule() -> Bool { hasModule }
+    
+    func startModule() {
+        moduleApi?.startModule()
+        moduleManager?.updateProviders(module)
+    }
+}