+++ AffiseAttributionLib/Classes/modules/advertising/AffiseAdvertisingApi.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+
+@objc
+public protocol AffiseAdvertisingApi: AffiseModuleApi {
+    func startModule()
+}