+++ AffiseAttributionLib/Classes/modules/attribution/AffiseAttributionModuleApi.swift
@@ -6,13 +6,6 @@ public protocol AffiseAttributionModuleApi : AnyObject {
      * Get module status
      */
     func getStatus(_ module: AffiseModules, _ onComplete: @escaping OnKeyValueCallback)
-    
-    /**
-    * Manual module start
-    */
-    @available(*, deprecated, message: "Will be removed")
-    @discardableResult
-    func moduleStart(_ module: AffiseModules) -> Bool
 
     /**
      * Get installed modules