+++ AffiseInternal/Classes/ext/AffiseSettingsExt.swift
@@ -10,7 +10,9 @@ internal class Field {
     // static let AUTO_CATCHING_CLICK_EVENTS = "autoCatchingClickEvents"
     static let IS_PRODUCTION = "isProduction"
     // static let ENABLED_METRICS = "enabledMetrics"
-     static let DOMAIN = "domain"
+    static let DOMAIN = "domain"
+    static let CONFIG_STRINGS = "configStrings"
+    static let DISABLE_MODULES = "disableModules"
 }
 
 internal extension Dictionary where Key == String, Value == Any? {
@@ -21,6 +23,12 @@ internal extension Dictionary where Key == String, Value == Any? {
     
     func getBool(_ key: String) -> Bool? {
         return self[key] as? Bool
+    }
+    func getMap(_ key: String) -> [String:Any?]? {
+        return self[key] as? [String:Any?]
+    }
+    func getList(_ key: String) -> [Any?]? {
+        return self[key] as? [Any?]
     }
     
     func isValid() -> Bool {
@@ -35,6 +43,11 @@ internal extension Dictionary where Key == String, Value == Any? {
     func getAppIdAndSecretKey() -> (String, String) {
         return (self.getString(Field.AFFISE_APP_ID) ?? "", self.getString(Field.SECRET_KEY) ?? "")
     }
+}
+
+internal func toAffiseModules(_ values: [Any?]?) -> [AffiseModules]? {
+    guard let values = values as? [String] else { return nil }
+    return values.compactMap { AffiseModules.from($0) }
 }
 
 internal extension AffiseSettings {
@@ -57,6 +70,18 @@ internal extension AffiseSettings {
         
         if let value = data.getString(Field.APP_TOKEN) {
             _ = self.setAppToken(value)
+        }
+
+        // if let value = data.getMap(Field.CONFIG_STRINGS) {
+        //     for (k, v) in value {
+        //         guard let v = v else { continue }
+        //         guard let configKey = AffiseConfig.from(k) else { continue }
+        //         _ = self.setConfigValue(configKey, value)
+        //     }
+        // }
+
+        if let value = toAffiseModules(data.getList(Field.DISABLE_MODULES)) {
+            _ = self.setDisableModules(value)
         }
         
         return self